moved the PotentialEmployeesPageCardsBlock folder to the client folder, connected Prismic slice for the header
