+++ frontend/src/client/page/PotentialEmployeesPage/PotentialEmployeesPage.tsx
@@ -1,46 +1,21 @@
-import PotentialEmployeesPageCardsBlock from '../../../widgets/PotentialEmployeesPageCardsBlock/ui/PotentialEmployeesPageCardsBlock';
 import { SliceZone, usePrismicDocumentByUID } from '@prismicio/react';
 import PotentialEmployeesStartBlock from '../../widgets/potentialEmployeesStartBlock/ui/potentialEmployeesStartBlock';
+import PotentialEmployeesPageCardsBlock from '../../widgets/PotentialEmployeesPageCardsBlock/ui/PotentialEmployeesPageCardsBlock';
+
 const PotentialEmployeesPage = () => {
   const [document] = usePrismicDocumentByUID('pages', 'potentialemployees');
   console.log(document);
   return (
     <>
-      <div style={{marginTop: 100}}>
+      <div style={{ marginTop: 100 }}>
         <SliceZone
           slices={document?.data.body}
           components={{
             startblock: PotentialEmployeesStartBlock,
+            potentialemployeescardsblock: PotentialEmployeesPageCardsBlock,
           }}
         />
       </div>
-      <PotentialEmployeesPageCardsBlock
-        // title="Сотрудники"
-        // data={[
-        //   {
-        //     _id: '1',
-        //     name: 'Арсен Белеков',
-        //     photo: '',
-        //     profession: 'Графический Дизайнер',
-        //     age: 23,
-        //     country: 'Кыргызстан',
-        //     city: 'Бишкек',
-        //     education: 'Высшее',
-        //     experience: 4,
-        //   },
-        //   {
-        //     _id: '2',
-        //     name: 'Арсен Белеков',
-        //     photo: '',
-        //     profession: 'Графический Дизайнер',
-        //     age: 23,
-        //     country: 'Кыргызстан',
-        //     city: 'Бишкек',
-        //     education: 'Высшее',
-        //     experience: 4,
-        //   },
-        // ]}
-      />
     </>
   );
 };
