+++ frontend/src/client/widgets/PotentialEmployeesPageCardsBlock/images/location-icon.png
