+++ frontend/src/client/widgets/PotentialEmployeesPageCardsBlock/images/person-icon.png
