+++ frontend/src/client/widgets/PotentialEmployeesPageCardsBlock/ui/PotentialEmployeesPageCard.css
