+++ frontend/src/client/widgets/PotentialEmployeesPageCardsBlock/ui/PotentialEmployeesPageCardsBlock.css
