+++ frontend/src/client/widgets/PotentialEmployeesPageCardsBlock/ui/PotentialEmployeesPageCardsBlock.tsx
@@ -1,12 +1,22 @@
 import PotentialEmployeesPageCard from './PotentialEmployeesPageCard';
 import './PotentialEmployeesPageCardsBlock.css';
-import { useAppDispatch, useAppSelector } from '../../../app/store/hooks';
-import { selectUsers, selectUsersLoading } from '../../../admin/page/usersPanel/model/usersSlice';
+import { useAppDispatch, useAppSelector } from '../../../../app/store/hooks';
+import { selectUsers, selectUsersLoading } from '../../../../admin/page/usersPanel/model/usersSlice';
 import { useEffect } from 'react';
-import { getAllUser } from '../../../admin/page/usersPanel/api/usersThunk';
-import { Loader } from '../../../shared/loader';
+import { getAllUser } from '../../../../admin/page/usersPanel/api/usersThunk';
+import { Loader } from '../../../../shared/loader';
 
-const PotentialEmployeesPageCardsBlock = () => {
+export interface PotentialEmployeesCardsBlockTitleApiData {
+  primary: {
+    titletext: string;
+  };
+}
+
+interface Props {
+  slice: PotentialEmployeesCardsBlockTitleApiData;
+}
+
+const PotentialEmployeesPageCardsBlock: React.FC<Props> = ({ slice }) => {
   const dispatch = useAppDispatch();
   const users = useAppSelector(selectUsers);
   const isLoading = useAppSelector(selectUsersLoading);
@@ -25,7 +35,7 @@ const PotentialEmployeesPageCardsBlock = () => {
 
   return (
     <div className="PotentialEmployeesPageCardsBlock__container">
-      <h2 className="PotentialEmployeesPageCardsBlock__title">Сотрудники</h2>
+      <h2 className="PotentialEmployeesPageCardsBlock__title">{slice.primary.titletext}</h2>
       {users.length === 0 ? (
         <p className="PotentialEmployeesPageCardsBlock__subtitle-no-cards">Нет сотрудников для отображения</p>
       ) : (
